+++ examples/schunk.py
@@ -10,7 +10,7 @@ import blosc2
 
 nchunks = 10
 # Set the compression and decompression parameters
-cparams = {"compcode": blosc2.Codec.LZ4HC, "typesize": 4}
+cparams = {"codec": blosc2.Codec.LZ4HC, "typesize": 4}
 dparams = {}
 contiguous = True
 urlpath = "filename"