+++ tests/test_compress2.py
@@ -14,7 +14,7 @@ import blosc2
 @pytest.mark.parametrize(
     "obj, cparams, dparams",
     [
-        (numpy.random.randint(0, 10, 10), {"compcode": blosc2.Codec.LZ4, "clevel": 6}, {}),
+        (numpy.random.randint(0, 10, 10), {"codec": blosc2.Codec.LZ4, "clevel": 6}, {}),
         (
             numpy.arange(10, dtype="float32"),
             # Select an absolute precision of 10 bits in mantissa
@@ -34,7 +34,7 @@ import blosc2
             {"splitmode": blosc2.SplitMode.ALWAYS_SPLIT, "nthreads": 5, "typesize": 4},
             {"schunk": None},
         ),
-        (numpy.arange(45, dtype=numpy.float64), {"compcode": blosc2.Codec.LZ4HC, "typesize": 4}, {}),
+        (numpy.arange(45, dtype=numpy.float64), {"codec": blosc2.Codec.LZ4HC, "typesize": 4}, {}),
         (numpy.arange(50, dtype=numpy.int64), {"typesize": 4}, blosc2.dparams_dflts),
     ],
 )
@@ -61,7 +61,7 @@ def test_compress2_numpy(obj, cparams, dparams):
 @pytest.mark.parametrize(
     "nbytes, cparams, dparams",
     [
-        (7, {"compcode": blosc2.Codec.LZ4, "clevel": 6, "typesize": 1}, {}),
+        (7, {"codec": blosc2.Codec.LZ4, "clevel": 6, "typesize": 1}, {}),
         (641091, {"typesize": 1}, {"nthreads": 4}),
         (136, {"typesize": 1}, {}),
         (1231, {"typesize": 4}, blosc2.dparams_dflts),
@@ -85,7 +85,7 @@ def test_compress2(nbytes, cparams, dparams):
 
 @pytest.mark.parametrize(
     "object, cparams, dparams",
-    [(numpy.arange(0), {"compcode": blosc2.Codec.LZ4, "clevel": 6}, {}), (b"", {}, {"nthreads": 3})],
+    [(numpy.arange(0), {"codec": blosc2.Codec.LZ4, "clevel": 6}, {}), (b"", {}, {"nthreads": 3})],
 )
 def test_raise_error(object, cparams, dparams):
     c = blosc2.compress2(object, **cparams, **dparams)