+++ tests/test_schunk_constructor.py
@@ -15,10 +15,10 @@ import blosc2
 @pytest.mark.parametrize(
     "cparams, dparams, chunksize",
     [
-        ({"compcode": blosc2.Codec.LZ4, "clevel": 6}, {}, 40000),
+        ({"codec": blosc2.Codec.LZ4, "clevel": 6}, {}, 40000),
         ({}, {"nthreads": 4}, 20000),
         ({"splitmode": blosc2.SplitMode.ALWAYS_SPLIT, "nthreads": 5}, {"schunk": None}, 20000),
-        ({"compcode": blosc2.Codec.LZ4HC, "typesize": 4}, {}, 40000),
+        ({"codec": blosc2.Codec.LZ4HC, "typesize": 4}, {}, 40000),
     ],
 )
 def test_schunk_numpy(contiguous, urlpath, cparams, dparams, chunksize):
@@ -63,7 +63,7 @@ def test_schunk_numpy(contiguous, urlpath, cparams, dparams, chunksize):
 @pytest.mark.parametrize(
     " cparams, dparams, chunksize",
     [
-        ({"compcode": blosc2.Codec.LZ4, "clevel": 6, "typesize": 1}, {}, 500),
+        ({"codec": blosc2.Codec.LZ4, "clevel": 6, "typesize": 1}, {}, 500),
         ({"typesize": 1}, {"nthreads": 4}, 500),
         ({"typesize": 1}, {}, 1000),
         ({"typesize": 1}, blosc2.dparams_dflts, 1000),